requirements.txt
cupy_cuda117==10.6.0
imageio==2.31.1
matplotlib==3.7.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.4.0
psutil==5.9.5
pykitti==0.3.1
scipy==1.10.1
scikit-image==0.21.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
wandb==0.15.10
